Use NSMetadataQuery to find accessible documents
Per the [documentation for searching file metadata with NSMetadataQuery
](https://developers.apple.com/library/archive/documentation/Carbon/Conceptual/SpotlightQuery/Concepts/QueryingMetadata.html),
we're creating a live query and listening for QueryDidFinishGathering
and QueryDidUpdate, and theoretically updating based on that.

At present, however, this query fires QueryDidFinishGathering with no
results, and never fires QueryDidUpdate.

I've added a CloudDocuments entitlement because the query fails to start
without that, but Apple's [Synchronizing Documents in the iCloud Environment
](https://developer.apple.com/documentation/uikit/documents_data_and_pasteboard/synchronizing_documents_in_the_icloud_environment)
suggests that we may also need to choose a container… though we're not
storing anything to any containers, so I'm not sure if that's necessary.
The sample code there also does some futzing with
FileManager.default.url(forUbiquityContainerIdentifier:) but then just…
never seems to use the output for anything relevant to the query, so
_that_ shouldn't matter either?

Anyway. This is what I think _should_ work, but doesn't at the moment.