Configure iCloud container
According to Apple's documentation
(https://developer.apple.com/documentation/uikit/documents_data_and_pasteboard/synchronizing_documents_in_the_icloud_environment),
in addition to the CloudDocuments entitlement, we (maybe?) also need to
specify the containers we're using. That said, that documentation is
mostly about an app that provides its own Documents folder.

In addition, it seems we need to call
FileManager.default.url(forUbiquityContainerIdentifier:) to set up those
container directories. While this is described in [the discussion on the
documentation for url(forUbiquityContainerIdentifier:)
](https://developer.apple.com/documentation/foundation/filemanager/1411653-url),
it's not exactly described very clearly anywhere else.

None of this appears to fix the issue, mind you, but I'm adding this
here to refer to later.