Monitor imported items using NSFilePresenter
Thanks to [Alexander Blach of Textastic, I discovered NSFilePresenter](https://twitter.com/Lextar/status/1438454801394712576),
which appears to be the best (only?) way to monitor the file system for
changes in imported folders/files. This adds a simple implementation to
illustrate how that might work in comparison to NSMetadataQuery.

It appears, though, that this has been [buggy since at least 2014](https://stackoverflow.com/questions/12389726/nsfilepresenter-presentedsubitemdidappearaturl-method-never-gets-called/19343784#19343784),
and most of the methods relating to subitems rarely, if ever, get
called. The only one that seems to be called consistently (maybe even
_too_ much?) is presentedSubitemDidChange(at:).